+++ src/src/stores/edge.ts
@@ -1,6 +1,6 @@
 import { Edge } from "@xyflow/react";
 import { atom } from "jotai";
 
-import { initialEdges } from "@/constant/initial/edge";
+import { initialEdges } from "@/constant/initial-edge";
 
 export const edgesAtom = atom<Edge[]>(initialEdges);
