+++ src/src/stores/node.ts
@@ -1,6 +1,6 @@
 import { Node } from "@xyflow/react";
 import { atom } from "jotai";
 
-import { initialNodes } from "@/constant/initial/node";
+import { initialNodes } from "@/constant/initial-node";
 
 export const nodesAtom = atom<Node[]>(initialNodes);
